Added garage model and fixed vehicle maintenance odometer fields including related models)

+++ Models/ViewModels/UpcomingMaintenanceViewModel.cs
@@ -6,9 +6,9 @@ namespace EaziLease.Models.ViewModels
         public string VehicleId {get; set;} = string.Empty;
         public string? RegistrationNumber {get; set;} = string.Empty;
         public DateTime? NextMaintenanceDate {get; set;}
-        public int? NextMaintenanceMileage {get; set;}
+        public decimal? NextMaintenanceMileage {get; set;}
         public int DaysRemaining {get; set;}
-        public int? KmRemaining {get; set;}
+        public decimal? KmRemaining {get; set;}
         public MaintenanceType Type {get; set;}
         public string StatusDisplay => DaysRemaining < 0 ? "Overdue" :
                                        DaysRemaining <= 7 ? "Due soon" : "Upcoming";
